+++ BO4E-dotnet/meta/LenientConverters/LenientParsing.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace BO4E
+namespace BO4E.meta.LenientParsing
 {
     /// <summary>
     /// Passing LenientParsing flags to <see cref="BoMapper"/> allows you to map such JSONs that
@@ -35,13 +35,12 @@ namespace BO4E
         /// </summary>
         SetInitialDateIfNull = 8,
         /// <summary>
-        /// most lenient (all others)
-        /// </summary>
-        MOST_LENIENT = ~0,
-        /// <summary>
-        /// Set integer value instead of String wrong old values
+        /// Try to parse Strings as Integer if type doesn't fit
         /// </summary>
         StringToInt = 16,
-
+        /// <summary>
+        /// most lenient (all others)
+        /// </summary>
+        MOST_LENIENT = ~0
     }
 }
