+++ TestBO4E-dotnet-Extensions/TestEnergiemengeExtension.cs
@@ -7,6 +7,7 @@ using BO4E.BO;
 using BO4E.ENUM;
 using BO4E.Extensions.BusinessObjects;
 using BO4E.Extensions.BusinessObjects.Energiemenge;
+using BO4E.meta.LenientParsing;
 using Itenso.TimePeriod;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
