+++ TestBO4E-dotnet-Extensions/TestEnergiemengeExtensionCompleteness.cs
@@ -7,6 +7,7 @@ using BO4E.BO;
 using BO4E.COM;
 using BO4E.ENUM;
 using BO4E.Extensions.BusinessObjects.Energiemenge;
+using BO4E.meta.LenientParsing;
 using BO4E.Reporting;
 using Itenso.TimePeriod;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
